requirements.txt
# Core ML frameworks
tensorflow-macos==2.13.0
mediapipe==0.10.21

# Protobuf (compatible with MediaPipe and TensorFlow)
protobuf>=4.25.3,<5

# Numpy
numpy<=1.24.3

# OpenCV for video processing
opencv-python==4.7.0.72

# Progress bar
tqdm==4.65.0

# Machine learning utilities
scikit-learn==1.3.0

# Optional dependencies for plotting/debugging
matplotlib==3.9.4

# ----------------------------
# API dependencies (compatible with TF 2.13 typing-extensions)
# ----------------------------
fastapi==0.95.2
starlette==0.27.0
uvicorn==0.23.2
python-multipart==0.0.6
